requirements.txt
emoji==1.6.1
Flask==2.0.1
Flask-Compress==1.10.1
Flask-SQLAlchemy==2.5.1
folium==0.2.1
huggingface-hub==0.2.1
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
pandas==1.3.1
requests==2.26.0
selenium==3.141.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.12.5
beautifulsoup4==4.9.3
numpy==1.19.5
regex==2021.8.28
line-bot-sdk==2.0.1
ckip-transformers==0.2.7
